requirements.txt
langchain 
langchain_core
langchain_community
langchain_openai
langchain_text_splitters
langchain-google-genai
langgraph
python-dotenv
pypdf
faiss-cpu
transformers
pydantic